pythonDjango4redis/.requirements.txt
Cheetah==2.4.4
#-e git+https://github.com/django/django.git@12e9adcd71efa36234b697e47fbf59c8117e29a1#egg=Django-dev
Flask==0.10.1
GnuPGInterface==0.3.2
Jinja2==2.7.2
#Landscape-Client==12.05
M2Crypto==0.21.1
MarkupSafe==0.19
MySQL-python==1.2.5
PAM==0.4.2
PIL==1.1.7
PyChef==0.2.2
PyMySQL==0.5
PyYAML==3.10
South==0.8.4
Twisted-Core>=11.1.0
Twisted-Names>=11.1.0
Twisted-Web>=11.1.0
Werkzeug==0.9.4
anyjson==0.3.3
apt-xapian-index==0.44
argparse==1.2.1
boto==2.13.3
chardet==2.0.1
cloud-init==0.6.3
command-not-found==0.2.44
configobj==4.7.2
## FIXME: could not find svn URL in dependency_links for this package:
distribute>=0.6.24dev-r0
django-dbindexer==1.6.1
django-markdown-deux==1.0.4
#django-redis-engine>=0.1.0
git+https://github.com/MirkoRossini/django-redis-engine.git
django-redisboard>=0.2.7
django-socketio==0.3.6
djangotoolbox==1.6.2
euca2ools==2.0.0
gevent==1.0
gevent-socketio==0.2.1
gevent-websocket==0.2.3
greenlet==0.4.2
httplib2==0.7.2
itsdangerous==0.23
keyring==0.9.2
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
markdown2==2.2.1
mercurial==2.9.1
oauth==1.0.1
paramiko==1.7.7.1
pexpect==2.3
prettytable==0.5
pyOpenSSL==0.12
pyasn1==0.0.11a
pycrypto==2.4.1
pycurl==7.19.0
pymongo==2.6.3
pyserial==2.5
pysnmp>=4.1.12a
pysnmp-apps>=0.2.6a
pysnmp-mibs>=0.0.5a
python-apt>=0.8.3ubuntu7
python-debian==0.1.21ubuntu1
redis==2.4.0
scalarizr>=2.4.5
scalr-upd-client>=0.4.10
simplejson==2.3.2
sphinx-me==0.2.1
ufw==0.31.1-1
unattended-upgrades==0.1
virtualenv==1.11.4
wadllib==1.3.0
wsgiref==0.1.2
zope.interface==3.6.1
